// libs/constants/src/enums/map-sort-type.enum.ts
export enum MapSortType {
  DATE_RELEASED_NEWEST = 0,
  DATE_RELEASED_OLDEST = 1,
  DATE_CREATED_NEWEST = 2,
  DATE_CREATED_OLDEST = 3,
  ALPHABETICAL = 4,
  ALPHABETICAL_REVERSE = 5,
  TIER_LOWEST = 6,
  TIER_HIGHEST = 7,
  PLAYED_NEWEST = 8,
  PLAYED_OLDEST = 9,
  PB_NEWEST = 10,
  PB_OLDEST = 11,
  FAVORITED_MOST = 12,
  FAVORITED_LEAST = 13,
  SUBMISSION_CREATED_NEWEST = 14,
  SUBMISSION_CREATED_OLDEST = 15,
  SUBMISSION_UPDATED_NEWEST = 16,
  SUBMISSION_UPDATED_OLDEST = 17
}